requirements.txt
django
djangorestframework
environs
mysqlclient
django-filter
django-cors-headers